csv_enricher/utils/csv_search_tool.py
from crewai_tools import CSVSearchTool as BaseCSVSearchTool
from typing import Dict, Any, Optional
import pickle
import os

class CachedCSVSearchTool(BaseCSVSearchTool):
    """
    Extension of CSVSearchTool that supports caching embeddings
    """
    
    def __init__(self, *args, **kwargs):
        """
        Initialize the CachedCSVSearchTool
        
        Args:
            *args: Positional arguments to pass to the parent class
            **kwargs: Keyword arguments to pass to the parent class
        """
        super().__init__(*args, **kwargs)
        self.embeddings_data = None
        
        # Check if cached data is provided
        config = kwargs.get('config', {})
        if config.get('use_cached_data', False) and 'cached_data' in config:
            self.embeddings_data = config['cached_data']
            print("Using cached embeddings data")
    
    def _run(self, *args, **kwargs):
        """
        Run the tool with caching support
        
        Args:
            *args: Positional arguments to pass to the parent class
            **kwargs: Keyword arguments to pass to the parent class
            
        Returns:
            The result of the search
        """
        # If we have cached embeddings data, use it
        if self.embeddings_data is not None:
            # TODO: Use the cached embeddings data
            # This would require modifying the internal implementation of CSVSearchTool
            # For now, we'll just pass it through to the parent class
            pass
        
        # Call the parent class implementation
        return super()._run(*args, **kwargs)
    
    def get_embeddings_data(self) -> Optional[Dict[str, Any]]:
        """
        Get the embeddings data for caching
        
        Returns:
            The embeddings data, or None if not available
        """
        # In a real implementation, we would extract the embeddings data from the tool
        # For now, we'll return a placeholder
        return {
            "version": "1.0",
            "timestamp": os.path.getmtime(self.csv),
            "csv_path": self.csv,
            # Add actual embeddings data here
        } 