csv_enricher/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.28.0
python-multipart==0.0.6
pandas==2.1.1
crewai[tools]==0.102.0
anthropic>=0.5.0
ollama==0.4.7
langchain-anthropic>=0.1.0
langchain-ollama>=0.1.0
litellm>=1.0.0
python-dotenv==1.0.0
tqdm==4.66.1 